Update LeafletHtmlIcon.js
Fixed a problem caused by div alignment before leaflet's translation: only first block was OK, because it floated to 0,0 by default. 
Also implemented a few options like iconSize and iconAnchor (was listed in available options but it didn't work for me) and an options to add or overwrite some custom CSS to the div.